refactor: move SettingsPage to settings/index.tsx to resolve upstream merge conflict
Agent-Logs-Url: https://github.com/tamaygz/job-ops/sessions/db209ea8-96f7-467e-b1ca-866de661dace
